e2e, tests: moved the ippool status to a dedicated pkg
